refactor(dashswallet): extract shared contract-resolution into helpers/dashswallet
CodeRabbit: the fees and dexs adapters carried an identical copy of the registry
constants, chainConfig and getContracts, so every change had to be applied twice
and the two could drift. Move SWAP_EXECUTED, SWAP_VERSION_REGISTRY, the registry
ABI, chainConfig, the resolved chains list, start, and getContracts into
helpers/dashswallet.ts and import them from both adapters. Each adapter keeps
only its own fetch/aggregation and adapter export.

Behavior unchanged: resolution, legacy/fallback, dedup and observability logging
are identical; address parity with 097ede92 preserved (76 targets); ts-check
clean; M.2 simulation green (18/18) against the shared module.
